fix: back to sign in button to /login instead of /signin and moved AccountConfig.js -> /imports/api/auth/ to match directory structure
